moved schema and types to its own file
